requirements.txt
requests
bs4
pandas
schedule
lxml
python-dotenv